Changed from Y-up axis to Z-up axis
I built the 3D model with the intent of rendering with MonoGame. MonoGame uses the Y-axis as "up". 3D printers use the Z-axis as "up". I changed the way the model was built so that it used the Z-axis as "up" instead of the Y-axis. This makes it in the right orientation for 3D printing.