Converted models to binary STLs
I ran some speed tests. I generated my 2022 model with the same settings in binary and in ascii 5 times each. I timed the conversion from file to WakaModel in the WakaModel.LoadFromFile() function.

There was an average of 84.14686 ms for the ascii loading and an average of 2.55932 ms for the binary loading.

That is a 96.9585% speed increase!!!!